lib/param: Add hook that allows modification of default settings.
This is useful for reducing the amount of configuration necessary for
OpenChange.

The hook is ideally registered from a plugin initialization function,
so that it automatically gets used whenever the plugin is installed.

This makes it possible for plugins to e.g. extend the default value for
the list of enabled dcerpc endpoint services.

Like all our interfaces, callers are expected to use this API
responsibly. For example, OpenChange should only enable its DCE/RPC
interface if it has been provisioned.

Change-Id: Ic8bacdd8b4c92a2a4b97cfa1a50dc41365b78071
Signed-Off-By: Jelmer Vernooij <jelmer@samba.org>

Reviewed-by: Andrew Bartlett <abartlet@samba.org>