waf: add a cross-execute program that runs natively
This added script is for supporting self-testing of the cross-build
infrastructure using autobuild.

The script can be used as the --cross-execute parameter
to samba's configure process, but it actually runs the program
natively on the build machine. This can be done if the build is
actually a native build disguised as a cross-build, i.e. native
toolchain is being used.

Signed-off-by: Uri Simchoni <urisimchoni@gmail.com>
Reviewed-by: Andrew Bartlett <abartlet@samba.org>
Reviewed-by: Alexander Bokovoy <ab@samba.org>